refactor(create-meeting): modularize CreateMeeting into components and hooks
