+++ client/src/pages/CreateMeeting/utils/utils.js
@@ -1,0 +1,7 @@
+export const generateRoomId = () => {
+  return (
+    Math.random().toString(36).substring(2, 10) +
+    "-" +
+    Math.random().toString(36).substring(2, 6)
+  );
+};
